consolidated dummy data and seed files into the same directory labeled script. updated package.json seed script to reflect this change.
