+++ package.json
@@ -34,7 +34,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "seed": "node seed.js"
+    "seed": "node script/seed.js"
   },
   "eslintConfig": {
     "extends": [
