+++ script/seed.js
@@ -1,21 +1,47 @@
-// const { db, Item } = require('../server/db');
+const { db, Item, Barter } = require('../server/db');
+const { ItemList } = require('./testData');
 
-// async function seed() {
-//   await db.sync({ force: true });
-//   console.log('db synced!');
+async function seed() {
+  try {
+    await db.sync({ force: true }); // clears db and matches models to tables
+    console.log('db synced!');
 
-//   const items = await Promise.all([
-//     Item.create({
-//       name: 'bicycle',
-//       description: 'check out my cool bike',
-//     }),
-//     Item.create({
-//       name: 'camera',
-//       description: 'here is my neat camera',
-//     }),
-//   ]);
-//   await db.close();
-//   return items;
-// }
+    await Promise.all(
+      ItemList.map((item) => {
+        return Item.create(item);
+      })
+    );
+  } catch (error) {
+    console.error('syncing dummy data did not work', error);
+  }
+}
 
-// seed();
+module.exports = seed;
+
+/*
+ We've separated the `seed` function from the `runSeed` function.
+ This way we can isolate the error handling and exit trapping.
+ The `seed` function is concerned only with modifying the database.
+*/
+async function runSeed() {
+  console.log('seeding...');
+  try {
+    await seed();
+  } catch (err) {
+    console.error(err);
+    process.exitCode = 1;
+  } finally {
+    console.log('closing db connection');
+    await db.close();
+    console.log('db connection closed');
+  }
+}
+
+/*
+  Execute the `seed` function, IF we ran this module directly (`node seed`).
+  `Async` functions always return a promise, so we can use `catch` to handle
+  any errors that might occur inside of `seed`.
+*/
+if (module === require.main) {
+  runSeed();
+}
